+++ guillotina/behaviors/dublincore.py
@@ -27,48 +27,58 @@ class IDublinCore(Interface):
 
     title = schema.TextLine(
         title=u'Title',
-        description=u"The first unqualified Dublin Core 'Title' element value.")
+        description=u"The first unqualified Dublin Core 'Title' element value.",
+        required=False)
 
     description = schema.Text(
         title=u'Description',
-        description=u"The first unqualified Dublin Core 'Description' element value.")
+        description=u"The first unqualified Dublin Core 'Description' element value.",
+        required=False)
 
     creation_date = schema.Datetime(
         title=u'Creation Date',
         description=u"The date and time that an object is created. "
-                    u"\nThis is normally set automatically.")
+                    u"\nThis is normally set automatically.",
+        required=False)
 
     modification_date = schema.Datetime(
         title=u'Modification Date',
         description=u"The date and time that the object was last modified in a\n"
-                    u"meaningful way.")
+                    u"meaningful way.",
+        required=False)
 
     effective_date = schema.Datetime(
         title=u'Effective Date',
-        description=u"The date and time that an object should be published. ")
+        description=u"The date and time that an object should be published. ",
+        required=False)
 
     expiration_date = schema.Datetime(
         title=u'Expiration Date',
-        description=u"The date and time that the object should become unpublished.")
+        description=u"The date and time that the object should become unpublished.",
+        required=False)
 
     creators = schema.Tuple(
         title=u'Creators',
         description=u"The unqualified Dublin Core 'Creator' element values",
-        value_type=schema.TextLine())
+        value_type=schema.TextLine(),
+        required=False)
 
     tags = schema.Tuple(
         title=u'Tags',
         description=u"The unqualified Dublin Core 'Tags' element values",
-        value_type=schema.TextLine())
+        value_type=schema.TextLine(),
+        required=False)
 
     publisher = schema.Text(
         title=u'Publisher',
-        description=u"The first unqualified Dublin Core 'Publisher' element value.")
+        description=u"The first unqualified Dublin Core 'Publisher' element value.",
+        required=False)
 
     contributors = schema.Tuple(
         title=u'Contributors',
         description=u"The unqualified Dublin Core 'Contributor' element values",
-        value_type=schema.TextLine())
+        value_type=schema.TextLine(),
+        required=False)
 
 
 @configure.behavior(